S3-22 Split monolithic modules into focused single-responsibility files
The domain crate's lib.rs (728 lines) mixed error types, domain structs,
Store trait, utility functions, and the full MemoryStore implementation.
Split into error.rs, types.rs, store.rs, util.rs, and memory_store.rs
with lib.rs as a thin re-export facade. Eliminated the now_secs()
duplication between MemoryStore and FsStore by centralizing shared
utilities in util.rs as pub(crate).

The server crate's handlers.rs (604 lines) mixed bucket ops, object ops,
date formatting, range parsing, and metadata extraction. Split into
handlers/ directory with bucket.rs, object.rs, and format.rs.

main.rs (127 lines) mixed config, routing, middleware, tracing, and
server startup. Extracted tracing_init.rs, store.rs, and router.rs
leaving main.rs as a 27-line entry point.
